Stream trace events over TCP
Use a TCP socket as transport to stream the trace events back to the
kubectl-gadget client. This commit uses port-forwarding to reach the
gadget pods without having to create a service / ingress.

The advantages of this mechanism are:
- Calling exec() into the pod is not needed
- No needed to parse output of a command (post process logic)
- In the future we could expose this as a service to make it independent
from the API server

TODO: Authentication needs to be done, otherwise anyone with access to
the port will be able to use the full gadget tracer manager API.