integration: Make busyboxPodCommand return a command
busyboxPodCommand() was returning a string, having a quite duplicated
logic in all the tests. This commit makes busyboxPodCommand() return
a command directly, the description is always "Run test-pod" and the
expected output match was removed as it's not needed.

A new busyboxPodRepeatCommand() function was introduced to run a command
multiple times (useful to avoid having to repeat while thrue; .. in all
command). Additionally, some tests that were using different images were
updated to use busybox as well.